feat: interface improvement
