+++ helpers/geometry/math.go
@@ -1,0 +1,16 @@
+package geometry
+
+// Helper functions to calculate min and max
+func min(a, b int32) int32 {
+	if a < b {
+		return a
+	}
+	return b
+}
+
+func max(a, b int32) int32 {
+	if a > b {
+		return a
+	}
+	return b
+}
