+++ vars/config.go
@@ -21,9 +21,12 @@ type PlatformDetails struct {
 }
 
 type ScreenDetails struct {
-	Width       int32 `json:"width"`
-	Height      int32 `json:"height"`
-	AspectRatio string
+	Width            int32 `json:"width"`
+	Height           int32 `json:"height"`
+	AspectRatio      string
+	MaxLines         int
+	MaxListItemWidth int
+	MaxListItens     int
 }
 
 type ConfigDefinition struct {
@@ -31,6 +34,26 @@ type ConfigDefinition struct {
 	Control      map[string]string          `json:"control"`
 	Screen       ScreenDetails              `json:"screen"`
 	Repositories map[string]PlatformDetails `json:"repositories"`
+}
+
+func LoadConfig(configFile []byte) (*ConfigDefinition, error) {
+	var config ConfigDefinition
+	err := json.Unmarshal(configFile, &config)
+	if err != nil {
+		if config.Logs {
+			log.Fatalf("Erro ao fazer o parse do arquivo embutido: %v", err)
+		}
+		return nil, err
+	}
+
+	aspectRation := calculateAspectRatio(config.Screen.Width, config.Screen.Height)
+	config.Screen.AspectRatio = aspectRation
+
+	config.Screen.MaxLines = calculateMaxLines(aspectRation)
+	config.Screen.MaxListItens = calculateMaxListItens(aspectRation)
+	config.Screen.MaxListItemWidth = calculateMaxListItemWidth(aspectRation)
+
+	return &config, nil
 }
 
 func calculateAspectRatio(width, height int32) string {
@@ -59,16 +82,54 @@ func calculateAspectRatio(width, height int32) string {
 	return fmt.Sprintf("%.2f:1", ratio)
 }
 
-func LoadConfig(configFile []byte) (*ConfigDefinition, error) {
-	var config ConfigDefinition
-	err := json.Unmarshal(configFile, &config)
-	if err != nil {
-		if config.Logs {
-			log.Fatalf("Erro ao fazer o parse do arquivo embutido: %v", err)
-		}
-		return nil, err
+func calculateMaxLines(aspectRatio string) int {
+	aspectRatios := map[string]int{
+		"16:9": 15,
+		"4:3":  15,
+		"3:2":  15,
+		"21:9": 15,
+		"32:9": 15,
+		"1:0":  15,
 	}
 
-	config.Screen.AspectRatio = calculateAspectRatio(config.Screen.Width, config.Screen.Height)
-	return &config, nil
+	return aspectRatios[aspectRatio]
+}
+
+func calculateMaxLineWidth(aspectRatio string) int {
+	aspectRatios := map[string]int{
+		"16:9": 30,
+		"4:3":  25,
+		"3:2":  20,
+		"21:9": 60,
+		"32:9": 50,
+		"1:0":  10,
+	}
+
+	return aspectRatios[aspectRatio]
+}
+
+func calculateMaxListItemWidth(aspectRatio string) int {
+	aspectRatios := map[string]int{
+		"16:9": 30,
+		"4:3":  25,
+		"3:2":  20,
+		"21:9": 60,
+		"32:9": 50,
+		"1:0":  10,
+	}
+
+	return aspectRatios[aspectRatio]
+}
+
+func calculateMaxListItens(aspectRatio string) int {
+	aspectRatios := map[string]int{
+		"16:9": 10,
+		"4:3":  11,
+		"3:2":  15,
+		"21:9": 10,
+		"32:9": 10,
+		"1:0":  10,
+	}
+
+	return aspectRatios[aspectRatio]
 }
